refactor(prebuilt): moved room layout fetching to `react-native-room-kit` from `react-native-hms`
