+++ packages/react-native-hms/package.json
@@ -104,7 +104,6 @@
     "zustand": "^4.3.8"
   },
   "devDependencies": {
-    "@100mslive/types-prebuilt": "^0.8.0",
     "@commitlint/config-conventional": "^11.0.0",
     "@react-native-community/eslint-config": "^2.0.0",
     "@types/jest": "^26.0.0",
